+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 0.4.3 - 2017-05-26
 ### Added
 - lower/upper triangular matrix inverter
+- gauss inverter
 
 ## 0.4.2 - 2017-05-25
 ### Refactored